@@ -1,6 +1,11 @@
 {
   "name": "ngx-diagrams-workspace",
   "version": "2.6.4",
+  "author": {
+    "email": "daniel.netzer7@gmail.com",
+    "name": "Daniel Netzer"
+  },
+  "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -20,47 +25,47 @@
     "url": "https://github.com/DanielNetzer/ngx-diagrams.git"
   },
   "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
+    "@angular/animations": "~8.2.4",
+    "@angular/common": "~8.2.4",
+    "@angular/compiler": "~8.2.4",
+    "@angular/core": "~8.2.4",
+    "@angular/forms": "~8.2.4",
+    "@angular/platform-browser": "~8.2.4",
+    "@angular/platform-browser-dynamic": "~8.2.4",
+    "@angular/router": "~8.2.4",
     "core-js": "^2.5.4",
-    "ngx-markdown": "^7.1.5",
-    "rxjs": "~6.3.3",
+    "ngx-markdown": "^8.1.0",
+    "rxjs": "~6.5.3",
 		"dagre": "^0.8.4",
-    "zone.js": "~0.8.26"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.0",
-    "@angular/cli": "~7.3.8",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@babel/core": "^7.4.4",
-    "@commitlint/cli": "^7.5.2",
-    "@commitlint/config-conventional": "^7.5.0",
-    "@semantic-release/changelog": "^3.0.2",
-    "@semantic-release/commit-analyzer": "^6.1.0",
-    "@semantic-release/exec": "^3.3.2",
-    "@semantic-release/git": "^7.0.8",
-    "@semantic-release/npm": "^5.1.7",
-    "@storybook/addon-notes": "^5.0.11",
-    "@storybook/addon-storysource": "^5.0.11",
-    "@storybook/angular": "^5.0.11",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-		"@types/lodash": "^4.14.126",
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular-devkit/build-ng-packagr": "~0.803.2",
+    "@angular/cli": "~8.3.2",
+    "@angular/compiler-cli": "~8.2.4",
+    "@angular/language-service": "~8.2.4",
+    "@babel/core": "^7.5.5",
+    "@commitlint/cli": "^8.1.0",
+    "@commitlint/config-conventional": "^8.1.0",
+    "@semantic-release/changelog": "^3.0.4",
+    "@semantic-release/commit-analyzer": "^6.3.0",
+    "@semantic-release/exec": "^3.3.6",
+    "@semantic-release/git": "^7.0.16",
+    "@semantic-release/npm": "^5.1.15",
+    "@storybook/addon-notes": "^5.1.11",
+    "@storybook/addon-storysource": "^5.1.11",
+    "@storybook/angular": "^5.1.11",
+    "@types/jasmine": "~3.4.0",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~12.7.4",
+		"@types/lodash": "^4.14.138",
 		"@types/dagre" :"^0.7.42",
     "babel-loader": "^8.0.5",
-    "codelyzer": "~4.5.0",
-    "cz-customizable": "^6.0.0",
-    "git-cz": "^3.0.1",
-    "husky": "^2.1.0",
+    "codelyzer": "~5.1.0",
+    "cz-customizable": "^6.2.0",
+    "git-cz": "^3.2.1",
+    "husky": "^3.0.5",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.0.0",
@@ -68,17 +73,17 @@
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "lint-staged": "^8.1.6",
-    "ng-packagr": "^4.2.0",
-    "prettier": "^1.17.0",
+    "lint-staged": "^9.2.5",
+    "ng-packagr": "^5.5.0",
+    "prettier": "^1.18.2",
     "protractor": "~5.4.0",
-    "semantic-release": "^15.13.16",
-    "ts-node": "~7.0.0",
+    "semantic-release": "^15.13.24",
+    "ts-node": "~8.3.0",
     "tsickle": ">=0.34.0",
-    "tslib": "^1.9.0",
-    "tslint": "~5.11.0",
+    "tslib": "^1.10.0",
+    "tslint": "~5.19.0",
     "tslint-config-prettier": "^1.18.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   },
   "config": {
     "commitizen": {
